refactor: abstract away SrcInfo interface
